+++ app/components/Views/Settings/SecuritySettings/Sections/RevealPrivateKey/styles.ts
